refactor: share WebGL context manager
